@@ -1,29 +1,43 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+    id 'org.jetbrains.kotlin.android' version '1.8.22' // 添加 Kotlin 插件
+}
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '27.0.3'
+    namespace "com.yemajun.hao.browser"
+    compileSdkVersion 35
+
     defaultConfig {
         applicationId "com.yemajun.hao.browser"
-        minSdkVersion 15
-        targetSdkVersion 28
+        minSdkVersion 21
+        targetSdkVersion 35
         versionCode 2
         versionName "2020.0201"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0-alpha3'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.0'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.8.22'  // 添加 Kotlin 标准库
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
+    implementation 'androidx.core:core-ktx:1.6.0'  // 引入 AndroidX 支持库
+    implementation ("androidx.webkit:webkit:1.8.0")
+    // Add these ExoPlayer dependencies
+    implementation 'androidx.media3:media3-exoplayer:1.3.1'
+    implementation 'androidx.media3:media3-exoplayer-dash:1.3.1' // Example: if you need DASH support
+    implementation 'androidx.media3:media3-ui:1.3.1'
+    // For offline downloads, you might specifically need:
+    implementation 'androidx.media3:media3-database:1.3.1'
 }
